% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-STRComparison.R
\name{scores}
\alias{scores}
\alias{params}
\alias{queryProfiles}
\alias{referenceProfiles}
\alias{STRComparison-accessors}
\alias{scores,STRComparison-method}
\alias{params,STRComparison-method}
\alias{queryProfiles,STRComparison-method}
\alias{referenceProfiles,STRComparison-method}
\alias{summary,STRComparison-method}
\alias{show,STRComparison-method}
\title{Accessors for STRComparison objects}
\usage{
scores(x, ...)

params(x, ...)

queryProfiles(x, ...)

referenceProfiles(x, ...)

\S4method{scores}{STRComparison}(x, ...)

\S4method{params}{STRComparison}(x, ...)

\S4method{queryProfiles}{STRComparison}(x, ...)

\S4method{referenceProfiles}{STRComparison}(x, ...)

\S4method{summary}{STRComparison}(object, ...)

\S4method{show}{STRComparison}(object)
}
\arguments{
\item{x, object}{A \link{STRComparison} object.}

\item{...}{Ignored.}
}
\value{
\code{scores()} returns the pair-level \link[S4Vectors:DataFrame]{S4Vectors::DataFrame}, \code{summary()} the
per-query one, \code{queryProfiles()} and \code{referenceProfiles()} the
\link{STRProfiles} that were compared, and \code{params()} a list of the scoring and
threshold settings used.
}
\description{
Extract the parts of the \link{STRComparison} object returned by
\code{\link[=compareProfiles]{compareProfiles()}}.
}
\examples{
q <- readSTRProfiles(
    system.file("extdata", "ExampleSTR_long.csv", package = "STRprofilerR"),
    sampleCol = "Sample Name"
)
ref <- readSTRProfiles(
    system.file("extdata", "ExampleSTR_database.csv", package = "STRprofilerR"),
    sampleCol = "Sample Name"
)
cmp <- compareProfiles(q, ref)

cmp
summary(cmp)
head(as.data.frame(scores(cmp)))
params(cmp)$tanThreshold

}
\author{
Jared Andrews
}
