% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-STRProfiles.R
\name{alleles}
\alias{alleles}
\alias{alleles<-}
\alias{markers}
\alias{markerData}
\alias{markerData<-}
\alias{sampleData}
\alias{sampleData<-}
\alias{provenance}
\alias{STRProfiles-accessors}
\alias{alleles,STRProfiles-method}
\alias{alleles<-,STRProfiles-method}
\alias{markers,STRProfiles-method}
\alias{markerData,STRProfiles-method}
\alias{markerData<-,STRProfiles-method}
\alias{sampleData,STRProfiles-method}
\alias{sampleData<-,STRProfiles-method}
\alias{provenance,STRProfiles-method}
\title{Accessors for STRProfiles objects}
\usage{
alleles(x, ...)

alleles(x, ...) <- value

markers(x, ...)

markerData(x, ...)

markerData(x, ...) <- value

sampleData(x, ...)

sampleData(x, ...) <- value

provenance(x, ...)

\S4method{alleles}{STRProfiles}(x, ...)

\S4method{alleles}{STRProfiles}(x, ...) <- value

\S4method{markers}{STRProfiles}(x, ...)

\S4method{markerData}{STRProfiles}(x, ...)

\S4method{markerData}{STRProfiles}(x, ...) <- value

\S4method{sampleData}{STRProfiles}(x, ...)

\S4method{sampleData}{STRProfiles}(x, ...) <- value

\S4method{provenance}{STRProfiles}(x, ...)
}
\arguments{
\item{x}{A \link{STRProfiles} object.}

\item{...}{Ignored.}

\item{value}{Replacement value.}
}
\value{
\code{alleles()} returns a \link[S4Vectors:DataFrame]{S4Vectors::DataFrame} of \link[IRanges:CharacterList]{IRanges::CharacterList}
columns, one column per marker.

\code{markers()} returns a character vector of marker names.

\code{markerData()} and \code{sampleData()} return a \link[S4Vectors:DataFrame]{S4Vectors::DataFrame} of marker
and sample annotations respectively.

\code{provenance()} returns a list describing how the object was built.

The replacement forms return an updated \link{STRProfiles} object.
}
\description{
Extract and replace the parts of a \link{STRProfiles} object.
}
\examples{
f <- system.file("extdata", "ExampleSTR_long.csv", package = "STRprofilerR")
p <- readSTRProfiles(f, sampleCol = "Sample Name")

markers(p)
markerData(p)
alleles(p)[["marker1"]]
provenance(p)$timestamp

# Correct a marker classification by hand.
markerData(p)$class[markers(p) == "marker1"] <- "y-linked"
markerData(p)

}
\author{
Jared Andrews
}
