% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{writeSTRProfiles}
\alias{writeSTRProfiles}
\title{Write STR profiles to file}
\usage{
writeSTRProfiles(x, file, sampleCol = "Sample")
}
\arguments{
\item{x}{A \link{STRProfiles} object.}

\item{file}{Path to write to.}

\item{sampleCol}{Name to give the sample identifier column.}
}
\value{
Invisibly, the path written to.
}
\description{
Writes a \link{STRProfiles} object to a CSV in the wide layout that
\code{\link[=readSTRProfiles]{readSTRProfiles()}} reads back, with alleles collapsed to comma-separated
strings.
}
\examples{
db <- system.file("extdata", "ExampleSTR_database.csv", package = "STRprofilerR")
p <- readSTRProfiles(db, sampleCol = "Sample Name")

out <- file.path(tempdir(), "profiles.csv")
writeSTRProfiles(p, out)
head(read.csv(out, check.names = FALSE))

}
\author{
Jared Andrews
}
